move news-crawler into bundle
