+++ extensions/o-mate/src/news-crawler/.gitignore
@@ -1,0 +1,2 @@
+.DS_Store
+node_modules
